Generate Cross Sections Across Hydrographic Network

Generate Cross Sections Across Hydrographic Network

Usage

cut_cross_sections(
  net,
  crosswalk_id = NULL,
  cs_widths = 100,
  num = 10,
  smooth = TRUE,
  densify = 2,
  rm_self_intersect = TRUE,
  fix_braids = FALSE,
  braid_threshold = NULL,
  braid_method = "crosswalk_id",
  precision = 1,
  add = FALSE,
  verbose = TRUE
)

Arguments

net

Hydrographic LINESTRING Network

crosswalk_id

Unique Identifier in net

cs_widths

numeric, Bankfull Widths (length of cross sections for each net element)

num

numeric, Number of transects per Net element

smooth

logical, whether to smooth linestring geometries or not. Default is TRUE.

densify

numeric, how many times more points should be added to linestrings. Default is 2.

rm_self_intersect

logical, whether to remove self intersecting transect linestrings

fix_braids

logical, whether to fix braided transect lines or not. If TRUE, linestrings that are part of a braided network are augmented. Default is FALSE.

braid_threshold

numeric value, value of the total length of all flowlines in a braid. Only braids with total flowline lengths less than or equal to the threshold will be considered by function (i.e. determines that maximum braid size that fix_braid_transects() should operate on). Default is NULL, which will attempt to fix all the braid transects in the data

braid_method

The method to determine the geometries to cut. Options are "crosswalk_id", "component", or "neighbor". Default is "crosswalk_id"

precision

int, distance in meters. Only applicable when fix_braids = TRUE. This is the number of meters to approximate final cross section linestring length. Increasing this value will decrease runtime of cross section extension algorithm. Value you must be greater than 0. Default is 1

add

logical indicating whether to add original 'net' data to the outputted transect lines. Default is FALSE.

verbose

logical, whether to output messages or not. Default is TRUE, and messages will be given

Value

sf object of transect linestrings