Select standard cross section point columns Internal helper function for selecting cross section point columns aligning with standard data model for cross section points

Select standard cross section point columns Internal helper function for selecting cross section point columns aligning with standard data model for cross section points

Usage

select_cs_pts(cs_pts, crosswalk_id = NULL)

Arguments

cs_pts

dataframe, tibble, or sf dataframe

crosswalk_id

character, unique ID column

Value

dataframe, tibble, or sf dataframe with only relavent cross section point columns