Plots an X-Y scatter plot of cross section points

Plots an X-Y scatter plot of cross section points

Usage

plot_cs_pts(
  cs_pts,
  crosswalk_id = NULL,
  x = "pt_id",
  y = "Z",
  color = NULL,
  size = 1,
  grid = FALSE,
  scales = "free_y"
)

Arguments

cs_pts

data.frame of cross section points with columns hy_id, cs_id and columns for X and Y axises (i.e. "pt_id", "Z")

crosswalk_id

unique ID column name

x

character name of column in cs_pts to use for X axis

y

character name of column in cs_pts to use for Y axis

color

character name of column in cs_pts to color points on plot

size

numeric, size of the cs points, default is 1

grid

logical, if TRUE then use facet_grid, otherwise use facet_wrap. Default is FALSE (uses facet_wrap)

scales

Should scales be fixed ("fixed", the default), free ("free"), or free in one dimension ("free_x", "free_y")?

Value

ggplot2 object